+++ packages/astro-template/prettier.config.cjs
@@ -1,4 +1,4 @@
-module.exports = require('@sumup/foundry/prettier')({
+module.exports = require('@sumup-oss/foundry/prettier')({
   plugins: ['prettier-plugin-astro'],
   overrides: [
     {