+++ packages/circuit-ui/components/Button/base.tsx
@@ -127,6 +127,7 @@ export function createButtonComponent<Props>(
   // TODO: Refactor to `mapClassName` once the deprecations have been removed.
   mapProps: (props: Props) => CreateButtonComponentProps,
 ) {
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
   const Button = forwardRef<any, Props>((props, ref) => {
     const {
       children,
@@ -144,7 +145,8 @@ export function createButtonComponent<Props>(
       ...sharedProps
     } = mapProps(props);
 
-    const { Link } = useComponents();
+    const components = useComponents();
+    const Link = components.Link as AsPropType;
 
     const isLink = Boolean(sharedProps.href);